Fix duplicate package declaration issue (#66)
If the package ca-certificates has already been declared someplace else in the codebase that uses ca_cert then this patch will prevent ca_cert from causing a duplicate declaration error.

Of course, one could argue that the user of the ca_cert module should just pass $install_package as false, and you may be right in that. But I see no harm in applying this patch.